refactoring, add munit to frontend, add inlineInput
