+++ frontend-laminar/src/main/scala/EditElement.scala
@@ -1,5 +1,5 @@
-import com.raquo.laminar.api.L.{*, given}
 import com.raquo.laminar.api.L
+import com.raquo.laminar.api.L.*
 object EditElement {
 
   def render(mods: Modifier[TextArea]*): HtmlElement =
