+++ frontend-laminar/src/main/scala/Router0.scala
@@ -1,8 +1,8 @@
-import com.raquo.laminar.api.L.{*, given}
-import com.raquo.waypoint.*
-import org.scalajs.dom
 import com.github.plokhotnyuk.jsoniter_scala.core.*
 import com.github.plokhotnyuk.jsoniter_scala.macros.*
+import com.raquo.laminar.api.L.*
+import com.raquo.waypoint.*
+import org.scalajs.dom
 
 object Router0 {
