+++ frontend-laminar/src/main/scala/ViewElement.scala
@@ -1,62 +1,68 @@
-import App.textVar
+import cats.effect.IO
+import cats.effect.unsafe.implicits.global
 import com.monovore.decline.Help
 import com.raquo.airstream.core.Signal
+import com.raquo.airstream.eventbus.EventBus
+import com.raquo.airstream.ownership.Owner
 import com.raquo.laminar.DomApi
-import com.raquo.laminar.api.L.{*, given}
-import com.raquo.laminar.api.L
-import ru.d10xa.jsonlogviewer.Application
-import ru.d10xa.jsonlogviewer.JsonDetector
-import ru.d10xa.jsonlogviewer.JsonPrefixPostfix
-import ru.d10xa.jsonlogviewer.JsonLogLineParser
-import ru.d10xa.jsonlogviewer.TimestampFilter
+import com.raquo.laminar.api.L.*
 import ru.d10xa.jsonlogviewer.LogViewerStream
-import ru.d10xa.jsonlogviewer.LogLineFilter
-import fs2.*
-import ru.d10xa.jsonlogviewer.decline.Config.FormatIn
 import ru.d10xa.jsonlogviewer.decline.Config
-import ru.d10xa.jsonlogviewer.decline.Config
-import ru.d10xa.jsonlogviewer.decline.TimestampConfig
-import ru.d10xa.jsonlogviewer.decline.TimestampConfig
-import ru.d10xa.jsonlogviewer.formatout.ColorLineFormatter
-import ru.d10xa.jsonlogviewer.logfmt.LogfmtLogLineParser
+import ru.d10xa.jsonlogviewer.decline.yaml.ConfigYaml
+import ru.d10xa.jsonlogviewer.decline.yaml.Feed
 
-import scala.util.chaining.scalaUtilChainingOps
+import scala.concurrent.ExecutionContext.Implicits.global
+
 object ViewElement {
 
-  def runApp(
-    logLinesSignal: Signal[String],
-    configSignal: Signal[Either[Help, Config]]
-  ): Signal[HtmlElement] =
-    logLinesSignal.combineWith(configSignal).map {
-      case (string, Right(c)) =>
-        val jsonPrefixPostfix = JsonPrefixPostfix(JsonDetector())
-        val logLineParser = c.formatIn match
-          case Some(FormatIn.Logfmt) => LogfmtLogLineParser(c)
-          case Some(FormatIn.Json)   => JsonLogLineParser(c, jsonPrefixPostfix)
-          case other =>
-            throw new IllegalStateException(
-              s"Unsupported format-in value: $other"
-            )
-
-        fs2.Stream
-          .emits(string.split("\n"))
-          .filter(_.trim.nonEmpty)
-          .through(LogViewerStream.stream(c, logLineParser))
-          .map(Ansi2HtmlWithClasses.apply)
-          .toList
-          .mkString("<div>", "", "</div>")
-          .pipe(DomApi.unsafeParseHtmlString)
-          .pipe(foreignHtmlElement)
-      case (string, Left(help)) =>
-        pre(cls := "text-light", help.toString)
-    }
   def render(
     logLinesSignal: Signal[String],
     configSignal: Signal[Either[Help, Config]]
-  ): HtmlElement =
+  )(implicit owner: Owner): HtmlElement = {
+    val eventBus = new EventBus[HtmlElement]
+
+    logLinesSignal
+      .combineWith(configSignal)
+      .foreach {
+        case (string, Right(c)) =>
+          c.copy(configYaml =
+            Some(c.configYaml.getOrElse(
+              ConfigYaml(
+                filter = None,
+                formatIn = None,
+                commands = None,
+                feeds = Some(
+                  List(
+                    Feed(
+                      name = "inlineInput",
+                      commands = List.empty,
+                      inlineInput = Some(string),
+                      filter = c.filter,
+                      formatIn = c.formatIn
+                    )
+                  )
+                )
+              )
+            ))
+          )
+          val stream: fs2.Stream[IO, HtmlElement] = LogViewerStream
+            .stream[IO](c)
+            .map(Ansi2HtmlWithClasses.apply)
+            .map(_.mkString("<div>", "", "</div>"))
+            .map(DomApi.unsafeParseHtmlString)
+            .map(foreignHtmlElement)
+
+          stream.compile.lastOrError
+            .unsafeToFuture()
+            .foreach(eventBus.writer.onNext)
+
+        case (_, Left(help)) =>
+          eventBus.writer.onNext(pre(cls := "text-light", help.toString))
+      }(owner)
+
     pre(
       cls := "bg-dark font-monospace text-white",
-      child <-- runApp(logLinesSignal, configSignal)
+      child <-- eventBus.events
     )
-
+  }
 }
