+++ json-log-viewer/js/src/main/scala/ru/d10xa/jsonlogviewer/decline/yaml/ConfigYamlLoaderImpl.scala
@@ -1,0 +1,8 @@
+package ru.d10xa.jsonlogviewer.decline.yaml
+
+import cats.data.ValidatedNel
+
+class ConfigYamlLoaderImpl extends ConfigYamlLoader {
+
+  override def parseYamlFile(content: String): ValidatedNel[String, ConfigYaml] = ???
+}
