+++ json-log-viewer/js/src/main/scala/ru/d10xa/jsonlogviewer/shell/ShellImpl.scala
@@ -1,9 +1,6 @@
 package ru.d10xa.jsonlogviewer.shell
 
-import cats.effect.*
 import fs2.*
-
-import java.io.*
 
 class ShellImpl[F[_]] extends Shell[F] {
