+++ json-log-viewer/jvm/src/main/scala/ru/d10xa/jsonlogviewer/decline/ConfigInitImpl.scala
@@ -2,15 +2,17 @@ package ru.d10xa.jsonlogviewer.decline
 
 import cats.data.Validated
 import cats.effect.IO
-import ru.d10xa.jsonlogviewer.ConfigYamlReader
 import ru.d10xa.jsonlogviewer.decline.Config.FormatIn
 import cats.syntax.all.*
 import ru.d10xa.jsonlogviewer.decline.yaml.ConfigYaml
+import ru.d10xa.jsonlogviewer.decline.yaml.ConfigYamlLoader
+import ru.d10xa.jsonlogviewer.decline.yaml.ConfigYamlLoaderImpl
+import ru.d10xa.jsonlogviewer.decline.yaml.ConfigYamlReader
 
 import java.io.File
 
 class ConfigInitImpl extends ConfigInit {
-
+  
   override def initConfig(c: Config): IO[Config] = {
     def findConfigFile(
       baseName: String,
