+++ json-log-viewer/shared/src/main/scala/ru/d10xa/jsonlogviewer/Application.scala
@@ -4,15 +4,9 @@ import cats.effect.*
 import com.monovore.decline.Opts
 import com.monovore.decline.effect.CommandIOApp
 import fs2.*
-import fs2.io.*
-import ru.d10xa.jsonlogviewer.decline.Config
-import ru.d10xa.jsonlogviewer.decline.Config.FormatIn
 import ru.d10xa.jsonlogviewer.decline.ConfigInit
 import ru.d10xa.jsonlogviewer.decline.ConfigInitImpl
 import ru.d10xa.jsonlogviewer.decline.DeclineOpts
-import ru.d10xa.jsonlogviewer.decline.yaml.ConfigYaml
-import ru.d10xa.jsonlogviewer.logfmt.LogfmtLogLineParser
-import ru.d10xa.jsonlogviewer.shell.ShellImpl
 
 object Application
   extends CommandIOApp(
