+++ json-log-viewer/shared/src/main/scala/ru/d10xa/jsonlogviewer/ParseResultKeys.scala
@@ -6,6 +6,7 @@ import HardcodedFieldNames.loggerNameFieldName
 import HardcodedFieldNames.threadNameFieldName
 import HardcodedFieldNames.stackTraceFieldName
 import ru.d10xa.jsonlogviewer.decline.Config
+
 class ParseResultKeys(config: Config) {
   def getByKey(
     parseResult: ParseResult,
