+++ json-log-viewer/shared/src/main/scala/ru/d10xa/jsonlogviewer/decline/FormatInValidator.scala
@@ -2,7 +2,6 @@ package ru.d10xa.jsonlogviewer.decline
 
 import cats.data.NonEmptyList
 import cats.data.Validated
-import cats.data.ValidatedNel
 import ru.d10xa.jsonlogviewer.decline.Config.FormatIn
 
 object FormatInValidator {
