+++ json-log-viewer/shared/src/main/scala/ru/d10xa/jsonlogviewer/decline/yaml/Feed.scala
@@ -6,6 +6,7 @@ import ru.d10xa.jsonlogviewer.query.QueryAST
 case class Feed(
   name: String,
   commands: List[String],
+  inlineInput: Option[String],
   filter: Option[QueryAST],
   formatIn: Option[FormatIn]
 )
