+++ json-log-viewer/shared/src/main/scala/ru/d10xa/jsonlogviewer/formatout/RawFormatter.scala
@@ -1,11 +1,8 @@
 package ru.d10xa.jsonlogviewer.formatout
 
-import fansi.ErrorMode.Strip
-import fansi.EscapeAttr
 import fansi.Str
 import ru.d10xa.jsonlogviewer.OutputLineFormatter
 import ru.d10xa.jsonlogviewer.ParseResult
-import ru.d10xa.jsonlogviewer.decline.Config
 
 class RawFormatter extends OutputLineFormatter:
   override def formatLine(p: ParseResult): Str =
