+++ json-log-viewer/shared/src/main/scala/ru/d10xa/jsonlogviewer/shell/Shell.scala
@@ -1,6 +1,5 @@
 package ru.d10xa.jsonlogviewer.shell
 import fs2.*
-import cats.effect.*
 
 trait Shell[F[_]] {
   def mergeCommands(commands: List[String]): Stream[F, String]
